+++ themes/book/app/root.tsx
@@ -84,7 +84,7 @@ function createSearch(index: MystSearchIndex): ISearch {
 }
 
 /*
- * Component that shows a "no CSS loaded" warning when a page 
+ * Component that shows a "no CSS loaded" warning when a page
  * loads without the built-in MyST stylesheet. This can happen on static builds
  * when the BASE_URL doesn't match the deployment base URL.
  */
@@ -109,7 +109,7 @@ function NoCSSWarning() {
           position: 'fixed',
           left: '0px',
           top: '0px',
-          width: '100vw',
+          width: '100%',
           height: '100vh',
           fontSize: '4rem',
           padding: '1rem',