+++ packages/common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @myst-theme/common
 
+## 1.1.1
+
+## 1.1.0
+
 ## 1.0.1
 
 ### Patch Changes