+++ packages/frontmatter/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @myst-theme/frontmatter
 
+## 1.1.1
+
+## 1.1.0
+
 ## 1.0.1
 
 ### Patch Changes