+++ packages/myst-demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # myst-demo
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [ae018ba]
+  - myst-to-react@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- Updated dependencies [75475f8]
+- Updated dependencies [c519f9b]
+- Updated dependencies [9bfb08e]
+  - myst-to-react@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes